Move config migrate command into CLI module
