+++ vulture_whitelist.py
@@ -130,7 +130,6 @@ _.validate_private_root  # unused method (src/mindroom/config/agent.py)
 _.validate_template_dir  # unused method (src/mindroom/config/agent.py)
 get_available_rooms  # unused function (src/mindroom/api/main.py)
 config_path_cmd  # unused function (src/mindroom/cli/config.py)
-config_migrate  # unused function (src/mindroom/cli/config.py)
 service_status  # unused function (src/mindroom/cli/service.py)
 _.validate_delegate_to  # unused method (src/mindroom/config/main.py)
 _.handle_startendtag  # unused method (src/mindroom/matrix/message_builder.py)
